Handle recent types & settings changes on Operation and Account
- make sure no extra field is used in the app
- use currency settings to display confirmed/not-confirmed operation
- add blockHeight fetch when syncing account
