+++ src/components/OperationsList/index.js
@@ -6,19 +6,17 @@ import moment from 'moment'
 import { connect } from 'react-redux'
 import { compose } from 'redux'
 import { translate } from 'react-i18next'
-import { getIconByCoinType } from '@ledgerhq/currencies/react'
 import {
   groupAccountOperationsByDay,
   groupAccountsOperationsByDay,
 } from '@ledgerhq/wallet-common/lib/helpers/account'
-import type { Account, Operation as OperationType } from '@ledgerhq/wallet-common/lib/types'
+
+import type { Account } from '@ledgerhq/wallet-common/lib/types'
 
 import noop from 'lodash/noop'
 import keyBy from 'lodash/keyBy'
 
-import { getMarketColor, rgba } from 'styles/helpers'
-
-import type { Settings, T } from 'types/common'
+import type { T } from 'types/common'
 
 import { MODAL_OPERATION_DETAILS } from 'config/constants'
 
@@ -27,17 +25,10 @@ import { openModal } from 'reducers/modals'
 import IconAngleDown from 'icons/AngleDown'
 
 import Box, { Card } from 'components/base/Box'
-import CounterValue from 'components/CounterValue'
-import FormattedVal from 'components/base/FormattedVal'
 import Text from 'components/base/Text'
 import Defer from 'components/base/Defer'
 
-import ConfirmationCheck from './ConfirmationCheck'
-
-const DATE_COL_SIZE = 100
-const ACCOUNT_COL_SIZE = 150
-const AMOUNT_COL_SIZE = 150
-const CONFIRMATION_COL_SIZE = 44
+import Operation from './Operation'
 
 const calendarOpts = {
   sameDay: 'LL – [Today]',
@@ -46,45 +37,6 @@ const calendarOpts = {
   lastWeek: 'LL',
   sameElse: 'LL',
 }
-
-const Day = styled(Text).attrs({
-  color: 'dark',
-  fontSize: 3,
-  ff: 'Open Sans',
-})`
-  letter-spacing: 0.3px;
-  text-transform: uppercase;
-`
-
-const Hour = styled(Day).attrs({
-  color: 'grey',
-})``
-
-const OperationRaw = styled(Box).attrs({
-  horizontal: true,
-  alignItems: 'center',
-})`
-  cursor: pointer;
-  border-bottom: 1px solid ${p => p.theme.colors.lightGrey};
-  height: 68px;
-
-  &:last-child {
-    border-bottom: 0;
-  }
-
-  &:hover {
-    background: ${p => rgba(p.theme.colors.wallet, 0.04)};
-  }
-`
-
-const Cell = styled(Box).attrs({
-  px: 4,
-  horizontal: true,
-  alignItems: 'center',
-})`
-  width: ${p => (p.size ? `${p.size}px` : '')};
-  overflow: ${p => (p.noOverflow ? 'hidden' : '')};
-`
 
 const ShowMore = styled(Box).attrs({
   horizontal: true,
@@ -102,141 +54,6 @@ const ShowMore = styled(Box).attrs({
   }
 `
 
-const AddressEllipsis = styled.div`
-  display: block;
-  flex-shrink: 1;
-  min-width: 20px;
-  overflow: hidden;
-  text-overflow: ellipsis;
-  white-space: nowrap;
-`
-
-const Address = ({ value }: { value: string }) => {
-  const addrSize = value.length / 2
-
-  const left = value.slice(0, 10)
-  const right = value.slice(-addrSize)
-  const middle = value.slice(10, -addrSize)
-
-  return (
-    <Box horizontal color="smoke" ff="Open Sans" fontSize={3}>
-      <div>{left}</div>
-      <AddressEllipsis>{middle}</AddressEllipsis>
-      <div>{right}</div>
-    </Box>
-  )
-}
-
-const Operation = ({
-  account,
-  minConfirmations,
-  onAccountClick,
-  onOperationClick,
-  t,
-  op,
-  withAccount,
-  marketIndicator,
-}: {
-  account: Account,
-  minConfirmations: number,
-  onAccountClick: Function,
-  onOperationClick: Function,
-  t: T,
-  op: OperationType,
-  withAccount?: boolean,
-  marketIndicator: string,
-}) => {
-  const { unit, currency } = account
-  const time = moment(op.date)
-  const Icon = getIconByCoinType(account.currency.coinType)
-  const isNegative = op.amount < 0
-  const type = !isNegative ? 'from' : 'to'
-
-  const marketColor = getMarketColor({
-    marketIndicator,
-    isNegative,
-  })
-
-  return (
-    <OperationRaw onClick={() => onOperationClick({ operation: op, account, type, marketColor })}>
-      <Cell size={CONFIRMATION_COL_SIZE} align="center" justify="flex-start">
-        <ConfirmationCheck
-          confirmations={op.confirmations}
-          marketColor={marketColor}
-          minConfirmations={minConfirmations}
-          t={t}
-          type={type}
-        />
-      </Cell>
-      <Cell size={DATE_COL_SIZE} justifyContent="space-between" px={3}>
-        <Box>
-          <Box ff="Open Sans|SemiBold" fontSize={3} color="smoke">
-            {t(`operationsList:${type}`)}
-          </Box>
-          <Hour>{time.format('HH:mm')}</Hour>
-        </Box>
-      </Cell>
-      {withAccount &&
-        account && (
-          <Cell
-            noOverflow
-            size={ACCOUNT_COL_SIZE}
-            horizontal
-            flow={2}
-            style={{ cursor: 'pointer' }}
-            onClick={e => {
-              e.stopPropagation()
-              onAccountClick(account)
-            }}
-          >
-            <Box
-              alignItems="center"
-              justifyContent="center"
-              style={{ color: account.currency.color }}
-            >
-              {Icon && <Icon size={16} />}
-            </Box>
-            <Box ff="Open Sans|SemiBold" fontSize={3} color="dark">
-              {account.name}
-            </Box>
-          </Cell>
-        )}
-      <Cell grow shrink style={{ display: 'block' }}>
-        <Address value={op.address} />
-      </Cell>
-      <Cell size={AMOUNT_COL_SIZE} justify="flex-end">
-        <Box alignItems="flex-end">
-          <FormattedVal
-            val={op.amount}
-            unit={unit}
-            showCode
-            fontSize={4}
-            alwaysShowSign
-            color={op.amount < 0 ? 'smoke' : undefined}
-          />
-          <CounterValue
-            color="grey"
-            fontSize={3}
-            date={time.toDate()}
-            ticker={currency.units[0].code}
-            value={op.amount}
-          />
-        </Box>
-      </Cell>
-    </OperationRaw>
-  )
-}
-
-Operation.defaultProps = {
-  onAccountClick: noop,
-  onOperationClick: noop,
-  withAccount: false,
-}
-
-const mapStateToProps = state => ({
-  settings: state.settings,
-})
-
 const mapDispatchToProps = {
   openModal,
 }
@@ -251,7 +68,6 @@ type Props = {
   withAccount?: boolean,
   nbToShow: number,
   title?: string,
-  settings: Settings,
 }
 
 export class OperationsList extends PureComponent<Props> {
@@ -271,7 +87,6 @@ export class OperationsList extends PureComponent<Props> {
       canShowMore,
       nbToShow,
       onAccountClick,
-      settings,
       t,
       title,
       withAccount,
@@ -295,7 +110,7 @@ export class OperationsList extends PureComponent<Props> {
               {title}
             </Text>
           )}
-          {groupedOperations.map(group => {
+          {groupedOperations.sections.map(group => {
             const d = moment(group.day)
             return (
               <Box flow={2} key={group.day.toISOString()}>
@@ -312,8 +127,6 @@ export class OperationsList extends PureComponent<Props> {
                       <Operation
                         account={account}
                         key={`${account.id}-${op.hash}`}
-                        marketIndicator={settings.marketIndicator}
-                        minConfirmations={account.minConfirmations}
                         onAccountClick={onAccountClick}
                         onOperationClick={this.handleClickOperation}
                         op={op}
@@ -338,4 +151,4 @@ export class OperationsList extends PureComponent<Props> {
   }
 }
 
-export default compose(translate(), connect(mapStateToProps, mapDispatchToProps))(OperationsList)
+export default compose(translate(), connect(null, mapDispatchToProps))(OperationsList)
