+++ src/helpers/btc.js
@@ -2,6 +2,7 @@
 
 import ledger from 'ledger-test-library'
 import bitcoin from 'bitcoinjs-lib'
+import axios from 'axios'
 import type { OperationRaw } from '@ledgerhq/wallet-common/lib/types'
 
 import groupBy from 'lodash/groupBy'
@@ -40,7 +41,7 @@ export function computeOperation(addresses: Array<string>, accountId: string) {
       confirmations: t.confirmations,
       date: t.received_at,
       accountId,
-      blockHeight: 0,
+      blockHeight: t.block.height,
     }
   }
 }
@@ -196,6 +197,11 @@ export async function getAccount({
               })
             : getAddress({ type: 'external', index: 0 })
 
+        // TODO: in the future, it should be done with the libc call
+        const {
+          data: { height: blockHeight },
+        } = await axios.get('https://api.ledgerwallet.com/blockchain/v2/btc_testnet/blocks/current')
+
         const account = {
           ...nextAddress,
           coinType,
@@ -204,6 +210,8 @@ export async function getAccount({
           balanceByDay: getBalanceByDay(operations),
           rootPath,
           operations,
+          blockTime: new Date(),
+          blockHeight,
         }
 
         onProgress({
